+++ just_audio/lib/just_audio.dart
@@ -109,6 +109,7 @@ class AudioPlayer {
   bool _canUseNetworkResourcesForLiveStreamingWhilePaused = false;
   double _preferredPeakBitRate = 0;
   bool _playInterrupted = false;
+  bool _platformLoading = false;
   AndroidAudioAttributes? _androidAudioAttributes;
   final bool _androidApplyAudioAttributes;
   final bool _handleAudioSessionActivation;
@@ -1127,6 +1128,7 @@ class AudioPlayer {
       {Completer<void>? playCompleter, bool force = false}) {
     if (_disposed) return null;
     if (!force && (active == _active)) return _durationFuture;
+    _platformLoading = active;
 
     // Warning! Tricky async code lies ahead.
     // (This should definitely be made less tricky)
@@ -1206,9 +1208,16 @@ class AudioPlayer {
             sequence![index].duration = duration;
           }
         }
+        if (_platformLoading &&
+            message.processingState != ProcessingStateMessage.idle) {
+          _platformLoading = false;
+        }
         final playbackEvent = PlaybackEvent(
-          processingState:
-              ProcessingState.values[message.processingState.index],
+          // The platform may emit an idle state while it's starting up which we
+          // override here.
+          processingState: _platformLoading
+              ? ProcessingState.loading
+              : ProcessingState.values[message.processingState.index],
           updateTime: message.updateTime,
           updatePosition: message.updatePosition,
           bufferedPosition: message.bufferedPosition,
@@ -1270,14 +1279,14 @@ class AudioPlayer {
 
       _platformValue = platform;
 
-      if (audioSource != null) {
-        _playbackEventSubject.add(_playbackEvent = _playbackEvent.copyWith(
-          updatePosition: position,
-          processingState: ProcessingState.loading,
-        ));
-      }
-
       if (active) {
+        if (audioSource != null) {
+          _playbackEventSubject.add(_playbackEvent = _playbackEvent.copyWith(
+            updatePosition: position,
+            processingState: ProcessingState.loading,
+          ));
+        }
+
         final automaticallyWaitsToMinimizeStalling =
             this.automaticallyWaitsToMinimizeStalling;
         final playing = this.playing;
@@ -1399,7 +1408,7 @@ class AudioPlayer {
 /// Captures the details of any error accessing, loading or playing an audio
 /// source, including an invalid or inaccessible URL, or an audio encoding that
 /// could not be understood.
-class PlayerException {
+class PlayerException implements Exception {
   /// On iOS and macOS, maps to `NSError.code`. On Android, maps to
   /// `ExoPlaybackException.type`. On Web, maps to `MediaError.code`.
   final int code;
@@ -1417,7 +1426,7 @@ class PlayerException {
 
 /// An error that occurs when one operation on the player has been interrupted
 /// (e.g. by another simultaneous operation).
-class PlayerInterruptedException {
+class PlayerInterruptedException implements Exception {
   final String? message;
 
   PlayerInterruptedException(this.message);
@@ -1553,10 +1562,11 @@ class PlayerState {
   String toString() => 'playing=$playing,processingState=$processingState';
 
   @override
-  int get hashCode => toString().hashCode;
+  int get hashCode => hashValues(playing, processingState);
 
   @override
-  bool operator ==(dynamic other) =>
+  bool operator ==(Object other) =>
+      other.runtimeType == runtimeType &&
       other is PlayerState &&
       other.playing == playing &&
       other.processingState == processingState;
@@ -1577,11 +1587,14 @@ class IcyInfo {
   String toString() => 'title=$title,url=$url';
 
   @override
-  int get hashCode => toString().hashCode;
+  int get hashCode => hashValues(title, url);
 
   @override
-  bool operator ==(dynamic other) =>
-      other is IcyInfo && other.toString() == toString();
+  bool operator ==(Object other) =>
+      other.runtimeType == runtimeType &&
+      other is IcyInfo &&
+      other.title == title &&
+      other.url == url;
 }
 
 class IcyHeaders {
@@ -1618,8 +1631,15 @@ class IcyHeaders {
   int get hashCode => toString().hashCode;
 
   @override
-  bool operator ==(dynamic other) =>
-      other is IcyHeaders && other.toString() == toString();
+  bool operator ==(Object other) =>
+      other.runtimeType == runtimeType &&
+      other is IcyHeaders &&
+      other.bitrate == bitrate &&
+      other.genre == genre &&
+      other.name == name &&
+      other.metadataInterval == metadataInterval &&
+      other.url == url &&
+      other.isPublic == isPublic;
 }
 
 class IcyMetadata {
@@ -1636,11 +1656,14 @@ class IcyMetadata {
   IcyMetadata({required this.info, required this.headers});
 
   @override
-  int get hashCode => info.hashCode ^ headers.hashCode;
+  int get hashCode => hashValues(info, headers);
 
   @override
-  bool operator ==(dynamic other) =>
-      other is IcyMetadata && other.info == info && other.headers == headers;
+  bool operator ==(Object other) =>
+      other.runtimeType == runtimeType &&
+      other is IcyMetadata &&
+      other.info == info &&
+      other.headers == headers;
 }
 
 /// Encapsulates the [sequence] and [currentIndex] state and ensures
@@ -2022,7 +2045,10 @@ abstract class AudioSource {
   int get hashCode => _id.hashCode;
 
   @override
-  bool operator ==(dynamic other) => other is AudioSource && other._id == _id;
+  bool operator ==(Object other) =>
+      other.runtimeType == runtimeType &&
+      other is AudioSource &&
+      other._id == _id;
 }
 
 /// An [AudioSource] that can appear in a sequence.
@@ -2744,13 +2770,15 @@ class LockCachingAudioSource extends StreamAudioSource {
       for (var cacheResponse in inProgressResponses) {
         if (_progress >= cacheResponse.end) {
           // We've received enough data to fulfill the byte range request.
-          cacheResponse.controller.add(
-              data.sublist(0, data.length - (_progress - cacheResponse.end)));
+          final subEnd = min(data.length,
+              max(0, data.length - (_progress - cacheResponse.end)));
+          cacheResponse.controller.add(data.sublist(0, subEnd));
           cacheResponse.controller.close();
         } else {
           cacheResponse.controller.add(data);
         }
       }
+      inProgressResponses.removeWhere((element) => element.controller.isClosed);
       if (_requests.isEmpty) return;
       // Prevent further data coming from the HTTP source until we have set up
       // an entry in inProgressResponses to continue receiving live HTTP data.
@@ -3517,6 +3545,12 @@ class AndroidEqualizerBand {
     }
   }
 
+  /// Restores the gain after reactivating.
+  Future<void> _restore() async {
+    await (await _player._platform).androidEqualizerBandSetGain(
+        AndroidEqualizerBandSetGainRequest(bandIndex: index, gain: gain));
+  }
+
   static AndroidEqualizerBand _fromMessage(
           AudioPlayer player, AndroidEqualizerBandMessage message) =>
       AndroidEqualizerBand._(
@@ -3546,6 +3580,13 @@ class AndroidEqualizerParameters {
     required this.bands,
   });
 
+  /// Restore platform state after reactivating.
+  Future<void> _restore() async {
+    for (var band in bands) {
+      await band._restore();
+    }
+  }
+
   static AndroidEqualizerParameters _fromMessage(
           AudioPlayer player, AndroidEqualizerParametersMessage message) =>
       AndroidEqualizerParameters(
@@ -3571,7 +3612,10 @@ class AndroidEqualizer extends AudioEffect with AndroidAudioEffect {
   @override
   Future<void> _activate() async {
     await super._activate();
-    if (_parametersCompleter.isCompleted) return;
+    if (_parametersCompleter.isCompleted) {
+      await (await parameters)._restore();
+      return;
+    }
     final response = await (await _player!._platform)
         .androidEqualizerGetParameters(AndroidEqualizerGetParametersRequest());
     _parameters =